fin depot et retrait et liste

+++ AppNatProjetSession/src/main/java/com/example/appnatprojetsession/Models/Banque.java
@@ -24,7 +24,7 @@ public class Banque extends Compte{
         super.setSoldeCompte(super.getSoldeCompte()+this.montantRemplissage);
     }
 
-    public void retraitGuichet(int montantRetrait){
+    public void retraitGuichet(double montantRetrait){
         super.setSoldeCompte(super.getSoldeCompte()- montantRetrait);
     }
